cleanup, fix styles
