+++ src/components/edtr-io/plugins/types/applet.tsx
@@ -9,7 +9,7 @@ import {
   HeaderInput,
   entityType,
 } from './common'
-import { RevisionHistory } from './helpers/revision-history'
+import { RevisionHistoryLoader } from './helpers/revision-history-loader'
 import { Settings } from './helpers/settings'
 import { useLoggedInData } from '@/contexts/logged-in-data-context'
 
@@ -41,7 +41,7 @@ function AppletTypeEditor(props: EditorPluginProps<typeof appletTypeState>) {
     <div>
       <div className="page-header">
         {props.renderIntoToolbar(
-          <RevisionHistory
+          <RevisionHistoryLoader
             id={props.state.id.value}
             currentRevision={props.state.revision.value}
             onSwitchRevision={props.state.replaceOwnState}
