+++ src/components/edtr-io/plugins/types/text-exercise.tsx
@@ -11,7 +11,7 @@ import {
   OptionalChild,
   entityType,
 } from './common'
-import { RevisionHistory } from './helpers/revision-history'
+import { RevisionHistoryLoader } from './helpers/revision-history-loader'
 import { useLoggedInData } from '@/contexts/logged-in-data-context'
 
 export const textExerciseTypeState = entityType(
@@ -49,7 +49,7 @@ export function TextExerciseTypeEditor(
   return (
     <article className="text-exercise">
       {props.renderIntoToolbar(
-        <RevisionHistory
+        <RevisionHistoryLoader
           id={props.state.id.value}
           currentRevision={props.state.revision.value}
           onSwitchRevision={props.state.replaceOwnState}
